Standardize file names in packages/jest-util (#7293)
* Rename files to follow facebook's conventions

* Rename files to follow facebook's conventions for test
