+++ packages/jest-util/src/ErrorWithStack.js
