+++ packages/jest-util/src/__tests__/fakeTimers.test.js
@@ -14,7 +14,7 @@ describe('FakeTimers', () => {
   let FakeTimers, moduleMocker, timerConfig;
 
   beforeEach(() => {
-    FakeTimers = require('../fake_timers').default;
+    FakeTimers = require('../FakeTimers').default;
     const mock = require('jest-mock');
     const global = vm.runInNewContext('this');
     moduleMocker = new mock.ModuleMocker(global);
