+++ packages/jest-util/src/__tests__/isInteractive.test.js
@@ -19,7 +19,7 @@ it('Returns true when running on interactive environment', () => {
   process.stdout.isTTY = true;
   process.env.TERM = 'xterm-256color';
 
-  const isInteractive = require('../is_interative').default;
+  const isInteractive = require('../isInteractive').default;
   expect(isInteractive).toBe(true);
 });
 
@@ -31,7 +31,7 @@ it('Returns false when running on a non-interactive environment', () => {
   jest.doMock('is-ci', () => true);
   process.stdout.isTTY = false;
   process.env.TERM = 'xterm-256color';
-  isInteractive = require('../is_interative').default;
+  isInteractive = require('../isInteractive').default;
   expect(isInteractive).toBe(expectedResult);
 
   // Test with is-ci being false and isTTY false
@@ -39,7 +39,7 @@ it('Returns false when running on a non-interactive environment', () => {
   jest.doMock('is-ci', () => false);
   process.stdout.isTTY = false;
   process.env.TERM = 'xterm-256color';
-  isInteractive = require('../is_interative').default;
+  isInteractive = require('../isInteractive').default;
   expect(isInteractive).toBe(expectedResult);
 
   // Test with is-ci being true and isTTY true
@@ -47,7 +47,7 @@ it('Returns false when running on a non-interactive environment', () => {
   jest.doMock('is-ci', () => true);
   process.stdout.isTTY = true;
   process.env.TERM = 'xterm-256color';
-  isInteractive = require('../is_interative').default;
+  isInteractive = require('../isInteractive').default;
   expect(isInteractive).toBe(expectedResult);
 
   // Test with dumb terminal
@@ -55,6 +55,6 @@ it('Returns false when running on a non-interactive environment', () => {
   jest.doMock('is-ci', () => false);
   process.stdout.isTTY = false;
   process.env.TERM = 'dumb';
-  isInteractive = require('../is_interative').default;
+  isInteractive = require('../isInteractive').default;
   expect(isInteractive).toBe(expectedResult);
 });
