+++ packages/jest-util/src/clearLine.js
