+++ packages/jest-util/src/convertDescriptorToString.js
