+++ packages/jest-util/src/getConsoleOutput.js
