+++ packages/jest-util/src/isInteractive.js
