+++ packages/jest-util/src/setGlobal.js
